requirements.txt
# Core OCR Engines
pytesseract==0.3.13
easyocr==1.7.1
paddleocr==2.7.0.3
paddlepaddle==2.5.2  # CPU version

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.0
numpy==1.26.4
pdf2image==1.17.0
PyMuPDF==1.23.26

# Utilities
tqdm==4.66.2
pyyaml==6.0.1
loguru==0.7.2

# Optional: GPU support (uncomment if you have CUDA)
# paddlepaddle-gpu==2.5.2
# torch torchvision  # For EasyOCR GPU
